Extract prepareData and getOptions to own files
